backend/requirements.txt
Flask-PyMongo
flask_jwt
flask-cors
